+++ app/Http/Controllers/API/V1/CategoryController.php
@@ -2,8 +2,8 @@
 
 namespace App\Http\Controllers\API\V1;
 
-use App\Http\Requests\StoreCategoryRequest;
-use App\Http\Requests\UpdateCategoryRequest;
+use App\Http\Requests\API\V1\StoreCategoryRequest;
+use App\Http\Requests\API\V1\UpdateCategoryRequest;
 use App\Models\Category;
 
 class CategoryController extends Controller
